implement AI confirmation and correction modal for voice input

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, Float
 from datetime import datetime
 
 from .database import Base
@@ -8,7 +8,7 @@ class Expense(Base):
     __tablename__ = "expenses"
 
     id = Column(Integer, primary_key=True, index=True)
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
     category = Column(String(100), nullable=False)
     description = Column(String(255), nullable=False)
     date = Column(DateTime, default=datetime.now, nullable=False)
@@ -18,7 +18,7 @@ class Income(Base):
     __tablename__ = "incomes"
 
     id = Column(Integer, primary_key=True, index=True)
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
     category = Column(String(100), nullable=False) 
     description = Column(String(255), nullable=True)
     income_date = Column(DateTime, default=datetime.now, nullable=False)
